LPS-105162 move 1_1_0 to 1_1_1
