+++ modules/dxp/apps/saml/saml-persistence-service/src/main/java/com/liferay/saml/persistence/internal/upgrade/v1_1_1/util/SamlSpSessionTable.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.saml.persistence.internal.upgrade.v1_1_0.util;
+package com.liferay.saml.persistence.internal.upgrade.v1_1_1.util;
 
 import java.sql.Types;
